Fix bug where after fork, same uuid is generated.
After a `fork()` the node id in each child processes is still the same. Since
the only other item that changes is the timestamp, then child processes could
generate the same uuid if they originated in the same original parent process.

This fixes #1 by detecting when the pid has changed and regenerating the node id
at that point.